Add files via upload
These two files together form a companion to the readings spreadsheet for the Solid Rock Hebrew Bible (see https://github.com/jjmccollum/solid-rock-hb). I beg readers' indulgence, as these files were not originally intended for publication, and consequently are very inconsistent in whether they address a matter, in how fully they reference relevant literature, how clearly my train of thought (such as it is) is expressed, and how egregious the typos are. Still, I have been encouraged to make this material available, and I do hope that it might in some measure be helpful for those seeking to research the text of the Hebrew Bible.